+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   - The previous behavior is preserved (the `.sql` suffix is now optional), so this is a new feature, not a breaking change.
   - Big thanks to @guspower who made big contributions to this feature.
 - Update ApexCharts to [v4.4.0](https://github.com/apexcharts/apexcharts.js/releases/tag/v4.4.0): fixes multiple small bugs in the chart component.
+- Add a new `auto_submit` parameter to the form component. When set to true, the form will be automatically submitted when the user changes any of its fields, and the page will be reloaded with the new value. The validation button is removed.
+  - This is useful to quickly create filters at the top of a dashboard or report page, that will be automatically applied when the user changes them.
 
 ## 0.32.1 (2025-01-03)